Added fan control, javascript strict mode
The fan has a single rule which is the state set to "auto" (match ac state) or "on" (always on).

The html page shows the current rule "auto" or "on", and when the fan is on, the
html page shows the state in black, gray otherwise.

Files:

lessmostat.py
- Added fan control, right now the fan can only be set to "on" or "auto".

lessmostat.html
- Added display of fan state and interaction.
- Added strict mode, fixed variable declarations.

lessmostat.css
- Added display of fan state.

readme.md
- Removed fan control todo

tiles.pjg
- Added web UI pic with fan control

Tests:
- Controlled fan on and auto with the html ui
  - verified when AC turns on and fan control is auto, the fan turns on.
  - verified when AC turns off and fan control is auto, the fan turns off.
  - verified when AC turns off and fan control is on, the fan is left on.
  - verified when AC turns on and fan control is on, the fan is left on.
  - verified fan control can be set on and off when ac is on or off.