Use location.hostname instead of hard-coding the MQTT broker IP
This should simplify setting up the scripts for other people, since it doesn't
require modifying the html if the MQTT broker and the webserver are the same
machine, which is the usual setup.

Same as before this change, lessmostat.cfg still needs to be changed for the
esp8266 to communicate with the MQTT broker.

Files:

lessmostat.html
- Changed to use location.hostname instead of a hard-coded ip for the MQTT broker
  IP/hostname. This still assumes the webserver and the MQTT broker are the same
  machine, but there's little that can be done if they are not, other than
  hardcoding whatever IP.

readme.md
- Updated instructions

Tests:
- Ran in the browser both testing with the IP address and with the host name of
  the webserver in the url, verified both are able to reach the MQTT broker
  listening in the webserver machine.