Added presets, fan utilization display, abstracted char symbol names.
Presets are stored with long press/right click, set with short press/left click.

A preset records the current fan and ac rules. When a preset is set, the html
sets the fan and the ac states to that of the preset rule.

Presets are part of the configuration, stored in the cfg file and published
when state information is requested.

Presets are stored on the thermostat, vs. locally in the browser via
localStorage, since that makes them the same across all devices, and it's also
convenient when used from a browser that erases history/cookies/local storage
(eg Firefox Focus).

Files:

lessmostat.css
- Changed humid to util
- Added presets

lessmostat.html
- Changed humid to util
- Added fan utilization calculation
- Added presets
- Changed requestFanToggle to requestFanState since presets
- Moved humidity with units on top of the current temperature, joined fan and
  ac utilization under the current temperature.
- Abstracted out symbol chars (fanChar, acChar, celsiusChar...)

lessmostat.py
- Added control/store_preset topic. This copies the current rules in the given
  preset index.

lessmostat.cfg
- Added presets

readme.md
- Added presets and utilization.

tiles.jpg
- Updated with presets and utilization.

Tests:
- Stored and set presets, verified ac and fan are set accordingly.
- Verified fan utilization makes sense.