Small fix for tapping on preset html children not working.
When a preset had html children (eg 1/2 symbol char with superscript html
formatting), tapping on the children would fail to find the preset index and
to set/store the preset.

This is because of using "target" in the event handler, which is the child but
not html the event was bound to and that contains the presetN class (where the
preset index is extracted from).

The solution is to use delegateTarget instead of target, which returns the html
element the event was bound to, as opposed to the element the event happens in.

Additionally, change all the presets contain the preset class (which has the
event bound to it), rather than be children of an html element with the preset
class.

Also used delegateTarget on the other event handlers out of caution for the same
reason.

Files:

lessmostat.html
- Changed events to use delegateTarget instead of target
- Added class preset to presetN elements, removed preset parent element

lessmostat.css
- Removed preset from presetN class hierarchy.

Tests:
- Pressed preset subelements, verified presets are set/stored when necessary.