Fixed fadein not affecting arrows, fixed preset vs. heat z-fight, display dynamic number of presets support
The fadein ui feedback on arrows stopped working when the event changed to use
the delegateTarget. This was because the arrows class was overriding the
parent's fadein opacity.

This change removes all opacity from the arrow classes, so the fadein on
delegateTarget works.

There also was a z-fight between .heat and .preset that caused taps to go to
.heat instead of to the preset.

This change defines the preset elements after the .heat element so .presets
have tap preference over .heat when they intersect.

In addition, displaying a variable number of presets is now supported, since the
preset HTML elements are now created dynamically. This is done in a similar
way the .number element is done, by having a ~full height sliver that gets
translated to the center and rotated appropriately rather than by guessing
hard-coded positions as it was done before.

Note that the number of presets is still currently fixed at however many the
thermostat has stored, since there's no UI to add/remove presets only to
set/store.

Files:

lessmostat.html
- Removed <sup> tag for 1/2 since it was increasing the glyph height and wasn't
  aesthetically pleasing
- Reordered presets wrt. the small square/heat elements so taps on the preset
  number go to the preset even if they overlap the corner of the heat square.
- Changed presets HTML elements to be created dynamically with the right
  angle of separation between them.
- Created as many preset HTML elements as presets stored in the thermostat.

lessmostat.css
- Generalized preset position by using the same scheme used in .number: a 50%
  translation plus a rotation.
- Minor cleanup, changed rotateZ with rotate for consistency, removed unused
  hover classes, especially the ones used for the arrow that were overriding the
  fadein opacity.

Tests:
- Pressed preset subelements, pressed arrows, verified presets are set/stored,
  ui is faded
- Verified 1/2 presets interact, display and fade properly.
- Hacked to display 4 presets, verified the dynamic preset display works.