Split into multiple .py files, added logging, mqtt exeption handling, deploy script
Using a single monolithic lessmostat.py script causes out of memory error from
the micropython interperter when too much code is added. This error was hard to
notice because it happens at boot and when webrepl reconnects the messages are
already gone (one workaround for that is to put a small sleep in main.py).
The symptom is that the script doesn't run and the exported functions are not
present, but re-importing again doesn't error, since it was already imported and
re-importing is ignored.

Micropython out of memory can be worked around by splitting into different files
which is what this commit does.

This also treats mqtt connection errors as soft errors and ignores them retrying
to connect later, rather than restarting the machine.

This also adds a deploy.py script that talks to the webrepl and deploys the
files and shows the console. This helps greatly with development.

Files:

deploy.py
- Script to deploy the micropython files to the device via webrepl, restart the
  device, pull and display the log, and display console output forever

readme.md
- Updated with the latest features.

upython/config.py
- Configuration file functions moved from lessmostat.py

upython/lessmostat.py
- Moved configuration functions to config.py
- Moved time and ntp functions to syncedtime.py
- Moved mqtt publishing functions to mqtt.py
- Removed prints and used the new logging.py module
- Added code to set fan and ac to known state at startup

upython/logging.py
- Logging functions moved from lessmostat.py
- Added support to restart the log file when over a given size

upython/main.py
- Added sleep before importing modules in order to have time to reconnect and
  see any import errors in the webrepl console (this also helps a bit with
  hysteresis in case the script were to restart frequently)
- Removed print functions, used new logging module
- Removed old comments about soft and hard errors, since soft errors (mqtt
  connection error, etc) are now handled and don't need a hard reset anymore

upython/mqtt.py
- Moved mqtt publishing functions from lessmostat.py
- Encapsulated some global state in a wrapper
- Added mqtt connection, message publishing
- Added code to handle MQTT connection errors and retry without restarting the
  system, as those errors can be frequent when the mqtt server or the connection
  are down

upython/syncedtime.py
- ntp and time functions moved from lessmostat.py

Tests:
- Verified micropython compiler out of memory no longer hit with multiple .py
  files
- Verified general operation though html works
- Verified deploy properly deploys the files, restarts and waits forever
  displaying the console
- Verified compile errors show on the deploy console
- Verified MQTT server down doesn't cause a hard reset and reconnects when the
  server is back